1.Fixed bug: On the GenoricRepository delete only if the exsits 2.Fixed bug: On MoveFileToNewLocation just rename the actual file and delete the json file 3.Fixed bug: Validate file name does not exsits on the folder to coy to on copy operation

+++ FolderContentManager/Repositories/GenericFolderContentRepository.cs
@@ -50,11 +50,13 @@ namespace FolderContentManager.Repositories
 
         public void Delete(string name, string path, FolderContentType type)
         {
+            if(!FileManager.Exists(CreateJsonPath(name, path, type))) return;
             FileManager.Delete(CreateJsonPath(name, path, type));
         }
 
         public void Delete(string fullPath)
         {
+            if (!FileManager.Exists(fullPath)) return;
             FileManager.Delete(fullPath);
         }
